refactor(#4861): extract the shared vector retrieval leg into MCPVectorLeg
